refactor: update deployment scripts
